% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst.R
\name{metadata_fst}
\alias{metadata_fst}
\title{Read metadata from a fst file}
\usage{
metadata_fst(path, old_format = FALSE)
}
\arguments{
\item{path}{path to fst file}

\item{old_format}{must be FALSE, the old fst file format is deprecated and can only be read and converted with fst package versions 0.8.0 to 0.8.10.}
}
\value{
Returns a list with meta information on the stored dataset in path. Has class fstmetadata.
}
\description{
Method for checking basic properties of the dataset stored in path,
after checking if the requested file exists.
}
