% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst.R
\name{read_fst}
\alias{read_fst}
\title{Read fst files}
\usage{
read_fst(
  path,
  columns = NULL,
  from = 1,
  to = NULL,
  as.data.table = FALSE,
  old_format = FALSE
)
}
\arguments{
\item{path}{path to fst file}

\item{columns}{Column names to read The default is to read all columns.}

\item{from}{Read data starting from this row number}

\item{to}{Read data up until this row number. The default is to read to the last row of the stored dataset.}

\item{as.data.table}{If TRUE, the result will be returned as a data.table object. Any keys set on dataset x before writing will be retained. This allows for storage of sorted datasets. This option requires data.table package to be installed.}

\item{old_format}{must be FALSE, the old fst file format is deprecated and can only be read and converted with fst package versions 0.8.0 to 0.8.10.}
}
\value{
A data frame (or data.table) with the selected columns and rows.
}
\description{
Reads data frames from fast-storage ('fst') file,
after checking if the requested file and columns exist before attempting to read.
}
