% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fst.R
\name{write_fst}
\alias{write_fst}
\title{Write fst files}
\usage{
write_fst(x, path, compress = 50, uniform_encoding = TRUE)
}
\arguments{
\item{x}{a data frame to write to disk}

\item{path}{path to fst file}

\item{compress}{value in the range 0 to 100, indicating the amount of compression to use. Lower values mean larger file sizes. The default compression is set to 50.}

\item{uniform_encoding}{If 'TRUE', all character vectors will be assumed to have elements with equal encoding. The encoding (latin1, UTF8 or native) of the first non-NA element will used as encoding for the whole column. This will be a correct assumption for most use cases. If 'uniform.encoding' is set to 'FALSE', no such assumption will be made and all elements will be converted to the same encoding. The latter is a relatively expensive operation and will reduce write performance for character columns.}
}
\description{
writes data frames into fast-storage ('fst') files.
}
